organized the code
