+++ src/Game.java
@@ -11,7 +11,7 @@ public class Game {
     }
 
     public void update(){
-    	bird.update();
+        bird.update();
 
         if (keyboard.isDown(KeyEvent.VK_SPACE)) {
             System.out.println("space is down!");
